Dependency Inversion of Game - Completed dependency inversion of Game. If you wish to maintain the new functionality use BasicGame.
